+++ slack-lint-checks/src/test/java/slack/lint/retrofit/RetrofitJarLoader.kt
@@ -6,6 +6,6 @@ import com.android.tools.lint.checks.infrastructure.TestFiles.LibraryReferenceTe
 import java.io.File
 import slack.lint.BaseSlackLintTest
 
-/** Loads the test Retrofit 2 jar from resources. */
-fun BaseSlackLintTest.retrofit2Jar() =
-  LibraryReferenceTestFile(File(javaClass.classLoader.getResource("retrofit-2.9.0.jar")!!.toURI()))
+/** Loads the test Retrofit 3 jar from resources. */
+fun BaseSlackLintTest.retrofit3Jar() =
+  LibraryReferenceTestFile(File(javaClass.classLoader.getResource("retrofit-3.0.0.jar")!!.toURI()))